tour_creation_service/requirements.txt
flask
geopy
googlemaps
networkx
dacite
python-google-places
